pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "crewai-goodmem"
description = "GoodMem RAG tools for CrewAI agents"
readme = "README.md"
license = { text = "Apache-2.0" }
authors = [{ name = "PAIR Systems Inc" }]
requires-python = ">=3.10,<3.14"
keywords = ["crewai", "goodmem", "rag", "tools", "agents", "memory"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Typing :: Typed",
]
dependencies = [
    "crewai>=1.14.3",
    "requests>=2.33.0,<3",
]
dynamic = ["version"]

[project.optional-dependencies]
dev = [
    "pytest>=8",
    "pytest-cov>=5",
    "ruff>=0.6",
    "mypy>=1.11",
    "python-dotenv>=1",
    "types-requests",
]

[project.urls]
Homepage = "https://github.com/PAIR-Systems-Inc/goodmem-crewai"
Source = "https://github.com/PAIR-Systems-Inc/goodmem-crewai"
Documentation = "https://docs.goodmem.com/integrations/crewai"
Issues = "https://github.com/PAIR-Systems-Inc/goodmem-crewai/issues"
Changelog = "https://github.com/PAIR-Systems-Inc/goodmem-crewai/blob/main/CHANGELOG.md"

[tool.hatch.version]
path = "src/crewai_goodmem/_version.py"

[tool.hatch.build.targets.wheel]
packages = ["src/crewai_goodmem"]

[tool.hatch.build.targets.sdist]
include = [
    "src/crewai_goodmem",
    "tests",
    "examples",
    "README.md",
    "LICENSE",
    "NOTICE",
    "CHANGELOG.md",
    "pyproject.toml",
]

[tool.ruff]
target-version = "py310"
src = ["src", "tests", "examples"]

[tool.ruff.lint]
extend-select = [
    "E",
    "F",
    "B",
    "S",
    "RUF",
    "N",
    "W",
    "I",
    "T",
    "PERF",
    "PIE",
    "TID",
    "RET",
    "UP006",
    "UP007",
    "UP035",
    "UP037",
    "UP045",
]
ignore = ["E501"]

[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = ["S101", "S105", "S106", "RUF012", "N818", "B017"]
"examples/**/*.py" = ["T201", "S110", "BLE001"]

[tool.ruff.lint.isort]
combine-as-imports = true
force-sort-within-sections = true
section-order = ["future", "standard-library", "third-party", "first-party", "local-folder"]
lines-after-imports = 2

[tool.mypy]
strict = true
python_version = "3.12"
plugins = ["pydantic.mypy"]
exclude = ["tests/", "examples/"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-q --strict-markers"
markers = [
    "e2e: live tests that require a running GoodMem server (set GOODMEM_BASE_URL to opt in)",
]
